+++ src/widgets/dialogs/RestoreBackupsDialog.hpp
@@ -7,6 +7,7 @@
 #include <QDialog>
 #include <QDialogButtonBox>
 #include <QLabel>
+#include <QPushButton>
 
 namespace chatterino {